Move FormToJSON to pkg/lib/web
