+++ pkg/lib/authflowclient/authflowclienthandlers/form_to_json.go
@@ -1,0 +1,7 @@
+package authflowclienthandlers
+
+import (
+	"github.com/authgear/authgear-server/pkg/lib/web"
+)
+
+var FormToJSON = web.FormToJSON
